Renamed package tree to planner
