+++ go.mod
@@ -3,7 +3,6 @@ module github.com/genjidb/genji
 go 1.14
 
 require (
-	github.com/davecgh/go-spew v1.1.1
 	github.com/dgraph-io/badger/v2 v2.0.3
 	github.com/stretchr/testify v1.5.1
 	go.etcd.io/bbolt v1.3.4
