+++ sql/planner/statement.go
@@ -1,4 +1,4 @@
-package tree
+package planner
 
 import (
 	"fmt"
