+++ sql/planner/tree_test.go
@@ -1,4 +1,4 @@
-package tree
+package planner
 
 import (
 	"testing"
